separated menu toolbar into component
