+++ client/src/components/Copyright.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import {Typography, Link} from "@material-ui/core"
+
+
+export default function Copyright() {
+    return (
+        <Typography variant="body2" color="textSecondary" align="center">
+            {'Copyright © '}
+            <Link color="inherit" href="/">
+                FleetSheets
+            </Link>{' '}
+            {new Date().getFullYear()}
+            {'.'}
+        </Typography>
+    );
+}
